Fix pipeline bootstrapping (#13)
The current bootstrapping of the HTTP pipeline isn't fully correct.
We're currently always setting up the pipeline using secure upgrade.
This means that we need to wait for the SSL handshake to conclude to
finish negotiating the protocol we're using. However, if we are setting
up an insecure channel (i.e. without TLS) we never add a TLS handler to
the pipeline, causing the protocol negotiation handler to wait forever
for the handshake to finish. This means that insecure H1 requests
currently don't work.

This PR changes this to set up the pipeline for H1 if TLS is not being
used. We could consider adding h2c support.