Add server listening address property in NIOHTTPServer (#32)
Motivation:

Currently, when specifying port 0 in `HTTPServerConfiguration`, there is
no API to determine what port the `NIOHTTPServer` has actually been
bound to.

Modifications:

- Added a `SocketAddress` type for representing a host address and port.
- Added an async `listeningAddress` property on `NIOHTTPServer` that
returns a `SocketAddress` once the server has started listening (or
throws an error if the address couldn't be bound to or if the server has
closed).
- Added an internal `State` type to `NIOHTTPServer` for representing the
idle, listening, and closed states and managing transitions between
them.

Result:

Users can determine the address and port the `NIOHTTPServer` is
listening at.